requirements.txt
aiohttp==3.6.2
astroid==2.3.3
async-timeout==3.0.1
attrs==19.3.0
chardet==3.0.4
discord.py==1.3.2
idna==2.9
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
multidict==4.7.5
pylint==2.4.4
rope==0.16.0
six==1.14.0
typed-ast==1.4.1
websockets==8.1
wrapt==1.11.2
yarl==1.4.2
